requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
data-importing-tool @ git+https://github.com/BeeGroup-cimne/beegroup_importing_tool@importing_tool
et-xmlfile==1.1.0
geographiclib==1.52
geopy==2.2.0
happybase==1.2.0
html5lib==1.1
idna==3.4
isodate==0.6.1
kafka-python==2.0.2
neo4j==4.4.1
numpy==1.23.4
openpyxl==3.0.9
pandas==1.5.1
ply==3.11
pycryptodome==3.14.1
pymongo==3.11.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.19.2
python-Levenshtein==0.12.2
python-slugify==5.0.2
pytz==2022.6
rdflib==6.2.0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
text-unidecode==1.3
thefuzz==0.19.0
thriftpy2==0.4.14
urllib3==1.26.12
webencodings==0.5.1
xlrd==2.0.1
